+++ Items/Accessories/SpaceStone.cs
@@ -55,24 +55,24 @@ You are immune to the Chaos State Debuff
             if (player.controlHook && rodCD == 0 && Main.myPlayer == player.whoAmI)
             {
                 Vector2 vector32;
-                vector32.X = (float)Main.mouseX + Main.screenPosition.X;
+                vector32.X = Main.mouseX + Main.screenPosition.X;
                 if (player.gravDir == 1f)
                 {
-                    vector32.Y = (float)Main.mouseY + Main.screenPosition.Y - (float)player.height;
+                    vector32.Y = Main.mouseY + Main.screenPosition.Y - player.height;
                 }
                 else
                 {
-                    vector32.Y = Main.screenPosition.Y + (float)Main.screenHeight - (float)Main.mouseY;
+                    vector32.Y = Main.screenPosition.Y + Main.screenHeight - Main.mouseY;
                 }
-                vector32.X -= (float)(player.width / 2);
-                if (vector32.X > 50f && vector32.X < (float)(Main.maxTilesX * 16 - 50) && vector32.Y > 50f && vector32.Y < (float)(Main.maxTilesY * 16 - 50))
+                vector32.X -= player.width / 2;
+                if (vector32.X > 50f && vector32.X < Main.maxTilesX * 16 - 50 && vector32.Y > 50f && vector32.Y < Main.maxTilesY * 16 - 50)
                 {
                     int num246 = (int)(vector32.X / 16f);
                     int num247 = (int)(vector32.Y / 16f);
-                    if ((Main.tile[num246, num247].wall != 87 || (double)num247 <= Main.worldSurface || NPC.downedPlantBoss) && !Collision.SolidCollision(vector32, player.width, player.height))
+                    if ((Main.tile[num246, num247].wall != 87 || num247 <= Main.worldSurface || NPC.downedPlantBoss) && !Collision.SolidCollision(vector32, player.width, player.height))
                     {
                         player.Teleport(vector32, 1, 0);
-                        NetMessage.SendData(65, -1, -1, null, 0, (float)player.whoAmI, vector32.X, vector32.Y, 1, 0, 0);
+                        NetMessage.SendData(65, -1, -1, null, 0, player.whoAmI, vector32.X, vector32.Y, 1, 0, 0);
                         
                         rodCD = 30;
                     }
