+++ Items/Projectiles/IndigoSolution.cs
@@ -31,7 +31,7 @@ namespace AAMod.Items.Projectiles
             int dustType = mod.DustType("IndigoSolution");
             if (projectile.owner == Main.myPlayer)
             {
-                Convert((int)(projectile.position.X + (float)(projectile.width / 2)) / 16, (int)(projectile.position.Y + (float)(projectile.height / 2)) / 16, 2);
+                Convert((int)(projectile.position.X + projectile.width / 2) / 16, (int)(projectile.position.Y + projectile.height / 2) / 16, 2);
             }
             if (projectile.timeLeft > 133)
             {
@@ -72,7 +72,7 @@ namespace AAMod.Items.Projectiles
             {
                 projectile.ai[0] += 1f;
             }
-            projectile.rotation += 0.3f * (float)projectile.direction;
+            projectile.rotation += 0.3f * projectile.direction;
         }
 
         public void Convert(int i, int j, int size = 4)
