+++ Items/Projectiles/Ryugen.cs
@@ -33,7 +33,7 @@ namespace AAMod.Items.Projectiles
         public override Color? GetAlpha(Color lightColor)
         {
             //return Color.White;
-            return new Color(0, 200, 0, 0) * (1f - (float)projectile.alpha / 255f);
+            return new Color(0, 200, 0, 0) * (1f - projectile.alpha / 255f);
         }
     }
 }
