+++ Items/Projectiles/Taserblast.cs
@@ -29,7 +29,7 @@ namespace AAMod.Items.Projectiles
 
         public override void AI()
         {
-            projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X) + 1.57f;
+            projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X) + 1.57f;
             projectile.localAI[0] += 1f;
             if (projectile.localAI[0] > 130f) //projectile time left before disappears
             {
