+++ Items/Vanity/Glitched/GlitchesHat.cs
@@ -31,7 +31,7 @@ namespace AAMod.Items.Vanity.Glitched
 		
 		public override void UpdateVanitySet(Player player)
 		{
-			int dust = Dust.NewDust(new Vector2((float) player.position.X-14 + (player.direction == -1f ? 40 : 0), (float) player.position.Y-0), 6, 6, 6, Main.rand.Next(-10,10)/5, Main.rand.Next(-10,10)/5, 200, default(Color), 2.0f);
+			int dust = Dust.NewDust(new Vector2(player.position.X - 14 + (player.direction == -1f ? 40 : 0), player.position.Y - 0), 6, 6, 6, Main.rand.Next(-10,10)/5, Main.rand.Next(-10,10)/5, 200, default(Color), 2.0f);
 			Main.dust[dust].noGravity = true;
 			Main.dust[dust].noLight = true;
         }
